Changed input type to double for increased precision.

+++ lib/lib_fun.h
@@ -10,6 +10,6 @@
 //   Wl: The total word length (number of bits) of the fixed-point representation.
 //   Fl: The fractional length (number of bits) of the fixed-point representation.
 //   y: An array of uint8_t to store the fixed-point representation (output).
-void flp_to_fxp(float x, uint8_t Wl, uint8_t Fl, uint8_t y[]);
+void flp_to_fxp(double x, uint8_t Wl, uint8_t Fl, uint8_t y[]);
 
 #endif // LIB_FUN_H
